Move turbine grid plotting to visualization moduel
